Use email address to store the unique identifier instead of the username
Username is sometimes visible to users, so a nicer identifier is better.
Also, to allow editing of remaining user properties, a "valid" email address is required.
When a user changes their email address or an authserver user gets removed
and a new one added with the same username, they will get locked out, because no two users
with the same username can exist. To allow admins to correct the problem (removing the offending user)
when they locked themselves out, an underscore is appended to the username (we try that 15 times and then give up)